# swagger.yaml
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: "MyCustomer API"
host: "dev.customerpay.me"
basePath: "/"
...
schemes:
- "https"
paths:
...
  /store/update/{storeId}:
    put:
      tags:
      - "Store"
      summary: "Update an existing Store"
      description: ""
      operationId: "updateStore"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "storeId"
        in: "path"
        description: "ID of store to return"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Store object that needs to be updated"
        required: true
        schema:
          $ref: "#/definitions/Store"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Store"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Store not found"
        "405":
          description: "Validation exception"
  /store/delete/{storeId}:
    delete:
      tags:
      - "Store"
      summary: "Deletes a store"
      description: ""
      operationId: "deleteStore"
      produces:
      - "application/json"
      parameters:
      - name: "storeId"
        in: "path"
        description: "Store id to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Store not found"
...
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "x-access-token"
    in: "header"
definitions:
...
  Store:
    type: "object"
    required:
    - "name"
    - "phone"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      phone:
        type: "string"
      address:
        type: "string"
      tagline:
        type: "string"
...